[refactor] : 패키지 계층 구조 변경
